app/auth-api/requirements.txt
fastapi[standard]
uvicorn
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
passlib[argon2]
pydantic
pydantic-settings
Authlib
pytest
pytest-asyncio
pytest-cov
httpx
aiokafka
itsdangerous